Continue development of flight software.
Integrate a PID controller in an attempt to automatically control the
height (Z). "live.py" now locks on to the visual fiducials around the
gate, determines the center of the gate in three dimensions and
rotations and attempts to bring the quadcopter up to the height of the
gate. Not yet finished tweaking the PID to get this to work. I fear
that the quadcopter drift in the other axes are too severe for this
simple approach to work as the first test. It looks like we will need
to integrate in the PIDs for X and Y first. After this working we will
have to add PIDs for the rotation. Once this all done the quadcopter
should be able to rise in place and point towards the gate.

PID code was hacked along the way and it needs to be cleaned up.

Update some of the documentation.